Derive `Hash` and `Debug` for `Flip`
In my hacking, it became inconvenient that I couldn't derive `Hash` on structs
that contained a `Flip` - specifically, I wanted to store resources in a
`HashMap<Box<dyn Hash + Eq + ResourceDescriptor>>`, but my `impl
ResourceDescriptor for (Path, Flip)` failed because `Flip` is not `Hash`. This
commit adds four (4) tokens, deriving `Hash` and `Debug` (because why not,
right?) on `Flip`, in addition to `Clone, Copy, PartialEq, Eq`, which were
already derived.